Time: 653 ms (50.40%), Space: 58 MB (48.55%) - LeetHub

+++ 1711-count-good-meals/1711-count-good-meals.cpp
@@ -16,6 +16,6 @@ public:
             }
             mpp[v[i]]++;
         }
-    return counter %(int)mod;
+    return counter %mod;
     }
 };
